make num_samples a variable.
Note: this variable defines how quickly we get a current and position.
This uses `analog_read` `num_samples` timesand this simply takes a bit.
Docu says: 1x `analog_read` takes about 0.1ms.

So this effectively steers the duration of each loop and thus it
steers if the `archive` is long enough or if it overflows.